Разбиваю на компоненты
