+++ package.json
@@ -25,7 +25,7 @@
     "eslint-config-htmlacademy": "^9.0.0",
     "html-webpack-plugin": "^5.5.1",
     "style-loader": "^3.3.2",
-    "webpack": "^5.79.0",
+    "webpack": "5.79.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.13.3"
   },
