+++ src/view/const.js
@@ -1,0 +1,1 @@
+export const NUMBER_OF_POINTS = 3;
