+++ src/view/filter-view.js
@@ -1,8 +1,8 @@
 import { createElement } from '../render.js';
 
 function createFilterTemplate() {
-  return (
-    `<form class="trip-filters" action="#" method="get">
+  return /*html*/`
+  <form class="trip-filters" action="#" method="get">
     <div class="trip-filters__filter">
       <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything">
       <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
@@ -24,8 +24,7 @@ function createFilterTemplate() {
     </div>
 
     <button class="visually-hidden" type="submit">Accept filter</button>
-  </form>`
-  );
+  </form>`;
 }
 
 export default class FilterView {
