+++ src/view/sort-view.js
@@ -1,8 +1,8 @@
 import { createElement } from '../render.js';
 
 function createSortTemplate() {
-  return (
-    `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
+  return /*html*/`
+    <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
     <div class="trip-sort__item  trip-sort__item--day">
       <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day">
       <label class="trip-sort__btn" for="sort-day">Day</label>
@@ -27,8 +27,7 @@ function createSortTemplate() {
       <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer" disabled>
       <label class="trip-sort__btn" for="sort-offer">Offers</label>
     </div>
-  </form>`
-  );
+  </form>`;
 }
 
 export default class SortView {
