+++ src/view/waypoint-view.js
@@ -1,7 +1,7 @@
 import {createElement} from '../render.js';
 
 function createwaypointTemplate() {
-  return (`
+  return /*html*/`
   <li class="trip-events__item">
     <div class="event">
       <time class="event__date" datetime="2019-03-18">MAR 18</time>
@@ -39,7 +39,7 @@ function createwaypointTemplate() {
       </button>
     </div>
   </li>
-`);
+`;
 }
 
 export default class waypointView {
